consolidate more stuff into array.rs
