+++ pg_analytics/src/tableam/scan.rs
@@ -11,7 +11,7 @@ use pgrx::*;
 use std::any::type_name;
 use std::sync::Arc;
 
-use crate::datafusion::datatype::GetDatum;
+use crate::datafusion::array::GetDatum;
 use crate::datafusion::stream::Stream;
 use crate::datafusion::table::DatafusionTable;
 use crate::errors::{NotFound, ParadeError};
